Created FacetDropdownBase as a base component for dropdown
